requirements.txt
flask
flask-cors
openai
PyPDF2
python-docx
openpyxl
pillow
pytesseract
